+++ project/projections.go
@@ -114,24 +114,24 @@ var TransverseMercator = &Projection{
 	},
 }
 
-// ScalarMercator converts from lon/lat float64 to x,y uint64.
+// ScalarMercator converts from lon/lat float64 to x,y uint32.
 // This is the same as Google's world coordinates.
 var ScalarMercator struct {
-	Level    uint64
-	ToPlanar func(g geo.Point, level ...uint64) (x, y uint64)
-	ToGeo    func(x, y uint64, level ...uint64) geo.Point
+	Level    uint32
+	ToPlanar func(g geo.Point, level ...uint32) (x, y uint32)
+	ToGeo    func(x, y uint32, level ...uint32) geo.Point
 }
 
 func init() {
 	ScalarMercator.Level = 31
-	ScalarMercator.ToPlanar = func(g geo.Point, level ...uint64) (x, y uint64) {
+	ScalarMercator.ToPlanar = func(g geo.Point, level ...uint32) (x, y uint32) {
 		l := ScalarMercator.Level
 		if len(level) != 0 {
 			l = level[0]
 		}
 		return mercator.ScalarProject(g.Lon(), g.Lat(), l)
 	}
-	ScalarMercator.ToGeo = func(x, y uint64, level ...uint64) geo.Point {
+	ScalarMercator.ToGeo = func(x, y uint32, level ...uint32) geo.Point {
 		l := ScalarMercator.Level
 		if len(level) != 0 {
 			l = level[0]
